OCE-303 Moved OCE entry files into a separate folder
